Updated netwokr test to run for fixed amount of time

+++ network_test/download_ubuntu.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 
+if [ $# -ne 1 ]
+then
+	echo "usage: $0 TIME"
+	exit 65 # bad args
+fi
+
+WGET_TIME=$1
+
 wget -a wget-ubuntu.log -b http://www.ubuntu.com/start-download?distro=desktop&bits=32&release=lts &
 
-sleep 1200
+sleep $WGET_TIME
 
 killall wget
 rm start-download*
